flake: fix Darwin build
None of this stuff is necessary these days. This will produce an
executable linking against a Nix store `libiconv`, but as the build
isn’t static to begin with that should be fine. If a static build
is required in the future, `pkgsStatic` can be used as it is in the
`nix-installer` flake.